Function that prints the binary representation of a number.

+++ 0x14-bit_manipulation/1-print_binary.c
@@ -11,7 +11,7 @@
  *
  * Return: returns the remainder
  */
-unsigned long int mod(unsigned long int n, int divisor)
+unsigned long int mod(unsigned long int n, unsigned long int divisor)
 {
 	while (n >= divisor)
 		n -= divisor;
